% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_optistock_app.R
\name{optistock_app}
\alias{optistock_app}
\title{Run Shiny app to create sandbox optistock CPF curves}
\usage{
optistock_app()
}
\value{
NULL. Opens and runs the Shiny application that comes with the
\code{optistock} package
}
\description{
This function will open a Shiny app where you can play around with parameters
to see how the resulting CPF curve will change.
}
