% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spp_examples.R
\name{spp_examples}
\alias{spp_examples}
\alias{walleye_example}
\alias{musky_example}
\alias{rainbow_trout_example}
\alias{chinook_example}
\title{Species examples from optistocking manuscript}
\usage{
walleye_example()

musky_example()

rainbow_trout_example()

chinook_example()
}
\value{
NULL. Opens an R script with an example
}
\description{
This family of functions opens R scripts to run the scenarios that are used
in the optistocking paper. Simply call the function to open the file that
contains R code for the species' of interest.
}
\examples{
\dontrun{
walleye_example()
}
}
